+++ src/app/admin/honeymoon/map/page.tsx
@@ -1,0 +1,8 @@
+'use client';
+
+import MapTab from '../MapTab';
+import { useHoneymoonApi } from '../HoneymoonContext';
+
+export default function HoneymoonMapPage() {
+    return <MapTab api={useHoneymoonApi()} />;
+}